+++ dms-application/src/main/kotlin/team/aliens/dms/domain/tag/spi/CommandTagPort.kt
@@ -2,7 +2,7 @@ package team.aliens.dms.domain.tag.spi
 
 import team.aliens.dms.domain.tag.model.StudentTag
 
-interface CommandStudentTagPort {
+interface CommandTagPort {
 
     fun saveAllStudentTags(studentTags: List<StudentTag>)
