--- dms-application/src/main/kotlin/team/aliens/dms/domain/tag/spi/StudentTagPort.kt
@@ -1,3 +1,0 @@
-package team.aliens.dms.domain.tag.spi
-
-interface StudentTagPort : CommandStudentTagPort
