+++ dms-application/src/main/kotlin/team/aliens/dms/domain/tag/spi/TagPort.kt
@@ -1,4 +1,5 @@
 package team.aliens.dms.domain.tag.spi
 
 interface TagPort :
-    QueryTagPort
+    QueryTagPort,
+    CommandTagPort
