--- dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/tag/StudentTagPersistenceAdapter.kt
@@ -1,22 +1,0 @@
-package team.aliens.dms.persistence.tag
-
-import org.springframework.stereotype.Component
-import team.aliens.dms.domain.tag.model.StudentTag
-import team.aliens.dms.domain.tag.spi.StudentTagPort
-import team.aliens.dms.persistence.tag.mapper.StudentTagMapper
-import team.aliens.dms.persistence.tag.repository.StudentTagJpaRepository
-
-@Component
-class StudentTagPersistenceAdapter(
-    private val studentTagRepository: StudentTagJpaRepository,
-    private val studentTagMapper: StudentTagMapper
-) : StudentTagPort {
-
-    override fun saveAllStudentTags(studentTags: List<StudentTag>) {
-        studentTagRepository.saveAll(
-            studentTags.map {
-                studentTagMapper.toEntity(it)
-            }
-        )
-    }
-}
