+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/tag/TagPersistenceAdapter.kt
@@ -3,9 +3,12 @@ package team.aliens.dms.persistence.tag
 import com.querydsl.jpa.impl.JPAQueryFactory
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
+import team.aliens.dms.domain.tag.model.StudentTag
 import team.aliens.dms.domain.tag.model.Tag
 import team.aliens.dms.domain.tag.spi.TagPort
+import team.aliens.dms.persistence.tag.mapper.StudentTagMapper
 import team.aliens.dms.persistence.tag.mapper.TagMapper
+import team.aliens.dms.persistence.tag.repository.StudentTagJpaRepository
 import team.aliens.dms.persistence.tag.repository.TagJpaRepository
 import java.util.UUID
 
@@ -13,6 +16,8 @@ import java.util.UUID
 class TagPersistenceAdapter(
     private val tagRepository: TagJpaRepository,
     private val tagMapper: TagMapper,
+    private val studentTagRepository: StudentTagJpaRepository,
+    private val studentTagMapper: StudentTagMapper,
     private val queryFactory: JPAQueryFactory
 ) : TagPort {
     override fun queryTagsBySchoolId(schoolId: UUID): List<Tag> {
@@ -25,4 +30,12 @@ class TagPersistenceAdapter(
     override fun queryTagById(tagId: UUID) = tagMapper.toDomain(
         tagRepository.findByIdOrNull(tagId)
     )
+
+    override fun saveAllStudentTags(studentTags: List<StudentTag>) {
+        studentTagRepository.saveAll(
+            studentTags.map {
+                studentTagMapper.toEntity(it)
+            }
+        )
+    }
 }
